biplot fa2 model

biplot fa2 model

biplot_fa2(
  model,
  predictions,
  one = -1,
  second = -1,
  fscore = 2,
  fmult = 10,
  alpha = 0.3,
  alpha_site = 0.5,
  alpha_ind = 0.5,
  subtitle = NULL,
  gen = NULL,
  size_ind_biplot = 3
)

Arguments

model

An asreml object with Factor analytic structure (fa2 or more)

predictions

dataframe with predicted values by site

one

c(1,-1)

second

c(1,-1)

fscore

score to visualize genotypes and avoid overlapping

fmult

in order to see the sites and genotypes in the same plot we have to play with this value

alpha

opacity segments

alpha_site

opacity sites

alpha_ind

opacity individuals

subtitle

text

gen

vector with genotypes to show (optional)

size_ind_biplot

size_ind_biplot

Value

list with 3 ggplots

Examples

# library(tidyverse)
# library(asreml)
# library(agridat)
# data(besag.met)
# dat <- besag.met
#
# dat <- dat %>% arrange(county)
# model <- asreml(fixed = yield ~ 1 + county,
#                 random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
#                 residual = ~ dsum(~ units | county),
#                 data = dat,
#                 na.action = list(x="include",y="include"))
#
# pp <- predict(model, classify = "county")$pvals

# sites
# p1 <- biplot_fa2(model,predictions = pp, fscore = 1, one = 1, second = 1, fmult = 5)$g1
# # sites scaled
# p2 <- biplot_fa2(model,predictions = pp, fscore = 1, one = 1, second = 1, fmult = 5)$g2
# # biplot
# p3 <- biplot_fa2(model,predictions = pp, fscore = 1, one = 1, second = 1, fmult = 5)$g3