Check Experimental Design

Check Experimental Design

check_design_MET(
  data = NULL,
  genotype = NULL,
  trial = NULL,
  traits = NULL,
  rep = NULL,
  block = NULL,
  row = NULL,
  col = NULL
)

Arguments

data

A data.frame in a wide format.

genotype

A character string indicating the column in data that contains genotypes.

trial

A character string indicating the column in data that contains trials.

traits

A character vector specifying the traits for which the models should be fitted.

rep

A character string indicating the column in data that contains replicates.

block

A character string indicating the column in data that contains sub blocks.

row

A character string indicating the column in data that contains the row coordinates

col

A character string indicating the column in data that contains the column coordinates.

Value

An object of class checkAgri, a list of data.frames.

Examples

# library(agridat)
# library(agriutilities)
# data(besag.met)
# dat <- besag.met
# results <- check_design_MET(
#   data = dat,
#   genotype = "gen",
#   trial = "county",
#   traits = c("yield"),
#   rep = "rep",
#   block = "block",
#   col = "col",
#   row = "row"
# )