Connectivity Matrix

This function generates a (n_trial x n_trial) matrix with the amount of genotypes shared between each pair of trial.

connectivity_matrix(
  data = NULL,
  genotype = "germplasmName",
  trial = "trial",
  response = NULL
)

Arguments

data

A data.frame in a wide format.

genotype

A character string indicating the column in data that contains genotypes.

trial

A character string indicating the column in data that contains trials.

response

A character string specifying the trait.

Value

matrix

Examples

# \donttest{
library(agridat)
data(besag.met)
dat <- besag.met
connectivity_matrix(
  data = dat,
  genotype = "gen",
  trial = "county",
  response = "yield"
)
#>    C1 C2 C3 C4 C5 C6
#> C1 64 64 64 64 64 64
#> C2 64 64 64 64 64 64
#> C3 64 64 64 64 64 64
#> C4 64 64 64 64 64 64
#> C5 64 64 64 64 64 64
#> C6 64 64 64 64 64 64
# }