covcor_heat {agriutilities}	R Documentation

Correlation Covariance Heatmap

Description

Correlation Covariance Heatmap

Usage

covcor_heat(
  matrix,
  corr = TRUE,
  size = 4,
  digits = 3,
  legend = c(0.6, 0.7),
  upper_tri = FALSE,
  reorder = FALSE
)

Arguments

matrix
A numeric matrix.

corr
A logical value indicating if the matrix is in a scaled form (TRUE by default, correlation matrix)

size
A numeric value to define the letter size.

digits
A numeric integer to define the number of digits to plot.

legend
the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)

upper_tri
A logical value to plot the Lower or Upper Triangular Part of the matrix. FALSE by default.

reorder
A logical value to reorder by a Hierarchical Clustering. FALSE by default.

Value

A ggplot object showing the upper triangular elements of the matrix.

Examples

library(agriutilities)
data(iris)
M <- cor(iris[, -5])
covcor_heat(matrix = M, corr = TRUE)

[Package agriutilities version 1.2.3 Index]