h_cullis {agriutilities}	R Documentation

Cullis heritability for lme4 models

Description

Cullis heritability for lme4 models

Usage

h_cullis(model, genotype, re_MME = FALSE)

Arguments

model
Object of class lmer.

genotype
A character string indicating the column in data that contains genotypes.

re_MME
A logical value to ask if we want to reconstruct the mixed models equations to estimate the Cullis heritability. (FALSE by default)

Value

A numerical value of the Cullis heritability estimate. If re_MME is TRUE, a list with matrices of the mixed models equations is returned.

Author(s)

Paul Schmidt, Johan Aparicio.

Examples

library(lme4)
library(agridat)
library(agriutilities)
dat <- john.alpha
g.ran <- lmer(
  formula = yield ~ rep + (1 | gen) + (1 | rep:block),
  data = dat
)
h_cullis(model = g.ran, genotype = "gen")

[Package agriutilities version 1.2.3 Index]