ic_reml_asr {agriutilities}	R Documentation

Find the AIC and BIC for a set of models fitted using asreml

Description

The function is a wrapper for 'icREML' function described in Verbyla (2019).

Usage

ic_reml_asr(fm, scale = 1)

Arguments

fm
A list of asreml fitted model objects

scale
A scalar to scale the variance matrix of the estimated fixed effects (to ensure numerical stability of a log-determinant). Default value is 1.

Value

A data frame. The data frame has the following components

model : the names of the models

loglik : the full log-likelihood for each model

p : the number of fixed effects parameters for each model

q : the number of (non-zero) variance parameters for each model.

b : the number of variance parameters that are fixed or on the boundary. These parameters are not counted in the AIC or BIC.

AIC : the AIC for each model

BIC : the BIC for each model

logdet : the log-determinant used in adjusting the residual log-likelihood for each model

Author(s)

Ari Verbyla (averbyla at avdataanalytics.com.au)

References

Verbyla, A. P. (2019). A note on model selection using information criteria for general linear models estimated using REML. Australian & New Zealand Journal of Statistics, 61(1), 39-50.

[Package agriutilities version 1.2.3 Index]