stability {agriutilities}	R Documentation

Stability Coefficients

Description

Stability Coefficients

Usage

stability(
  predictions = NULL,
  genotype = NULL,
  trial = NULL,
  response = NULL,
  best = "max"
)

Arguments

predictions
A data.frame with one value per GxE combination.

genotype
A character string indicating the column in predictions that contains genotypes.

trial
A character string indicating the column in predictions that contains trials.

response
A character string specifying the response variable.

best
A character string specifying how to define the best genotype by numeric value ("min", "max"). "max" by default.

Value

A data.frame with several stability measures. "superiority" (cultivar-superiority measure), "static" (Shukla's stability variance) and "wricke" (Wricke's ecovalence).

Examples

## Not run: 
library(agridat)
library(agriutilities)

data(besag.met)
dat <- besag.met
results <- check_design_met(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
met_results <- met_analysis(out, progress = FALSE)

head(
  stability(
    predictions = met_results$BLUPs_GxE,
    genotype = "genotype",
    trial = "trial",
    response = "predicted.value"
  )
)

## End(Not run)

[Package agriutilities version 1.2.3 Index]