Extract Variance-Covariance from ASReml-R

Extract Variance-Covariance from ASReml-R

Usage

extract_vcov(model = NULL, gen = "genotype", env = "trial", vc_model = "corv")

Arguments

model

ASReml object

gen

A character string indicating the column in data that contains genotypes.

env

A character string indicating the column in data that contains environments or trials.

vc_model

A character string indicating the variance-covariance fitted. Can be 'diag', 'corv', 'corh', 'corgv', 'fa1', 'fa2', 'fa3', 'fa4', 'corgh', 'us' or 'rr2'.

Value

An object with a list of:

VCOV

A matrix of the estimated variance-covariance between trials.

CORR

A n_trial x n_trial matrix with the correlation between trials.

vc_model

A character string indicating the variance-covariance fitted.

Examples

# \donttest{
library(agridat)
library(asreml)
#> Loading required package: Matrix
#> Warning: package 'Matrix' was built under R version 4.1.3
#> Online License checked out Fri Feb  3 09:35:45 2023
library(agriutilities)

data(besag.met)
dat <- besag.met
results <- check_design_met(
 data = dat,
 genotype = "gen",
 trial = "county",
 traits = c("yield"),
 rep = "rep",
 block = "block",
 col = "col",
 row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
met_results <- met_analysis(out)
#> Online License checked out Fri Feb  3 09:35:47 2023

# Extract VCOV
extract_vcov(model = met_results$met_models$yield,
            vc_model = "us"
            )
#> $VCOV
#>          C1       C2       C3       C4        C5       C6
#> C1 80.57078 27.35631 46.80920 31.67095  85.61470 31.17032
#> C2 27.35631 28.90184 26.44403 20.40997  27.99220 33.50946
#> C3 46.80920 26.44403 79.66149 45.84374  64.53892 19.81346
#> C4 31.67095 20.40997 45.84374 29.28131  40.79469 20.75609
#> C5 85.61470 27.99220 64.53892 40.79469 100.59937 27.13428
#> C6 31.17032 33.50946 19.81346 20.75609  27.13428 66.66439
#> 
#> $CORR
#>           C1        C2        C3        C4        C5        C6
#> C1 1.0000000 0.5668993 0.5842762 0.6520442 0.9509593 0.4253099
#> C2 0.5668993 1.0000000 0.5511131 0.7015917 0.5191307 0.7634097
#> C3 0.5842762 0.5511131 1.0000000 0.9492062 0.7209412 0.2718875
#> C4 0.6520442 0.7015917 0.9492062 1.0000000 0.7516415 0.4697896
#> C5 0.9509593 0.5191307 0.7209412 0.7516415 1.0000000 0.3313399
#> C6 0.4253099 0.7634097 0.2718875 0.4697896 0.3313399 1.0000000
#> 
#> $vc_model
#> [1] "us"
#> 
# }