Factor analytic variance models.

Summarises an asreml model in which factor analytic (FA) variance structures have been fitted. The function aids in interpreting FA models fitted to genotype by environment effects in the analysis of plant breeding multi-environment trial (MET) data.

Usage

fa.asreml(
  object,
  uniplot = F,
  uniplot.tol = 0.85,
  uniplot.cex = 0.5,
  trunc.char = c(1, 6),
  blups = TRUE,
  regplot = F,
  addedplot = F,
  g.list = NULL,
  heatmap = F,
  heatmap.ord = "asis",
  agnes.method = "average"
)

Arguments

object

An asreml object.

uniplot

if TRUE (the default), pairwise plots of FA loadings are produced for k > 1.

uniplot.tol

A numeric value between 0 and 1 that affects the display of pairs of loadings in the uniplots. Those pairs with a (scaled) euclidean distance greater than uniplot.tol are drawn with solid blue lines, otherwise they appear as red dashed lines.

uniplot.cex

Character expansion (relative) size for plotting (the default is 0.5).

trunc.char

A numeric vector of length 2 specifying the truncation of loading labels for use in plotting. The default (c(1,6)) is to use the first 6 characters of the environment factor levels, otherwise if NULL the complete level names are used.

blups

If TRUE (the default), summaries of estimates from the asreml object are produced. These include BLUPs of genotype by environment effects and genotype scores, and REML estimates of FA variance parameters.

regplot

If TRUE (the default), regression plots of genotype by environment effects against FA loadings are produced.

addedplot

If TRUE (the default), added variable plots of genotype by environment effects against FA loadings are produced.

g.list

A character vector comprising the subset of genotype levels to be used in the regression and added variable plots. The default (NULL) is to use the complete set of genotype levels.

heatmap

If TRUE (the default), a heatmap of the estimated FA correlation matrix is produced.

heatmap.ord

A character string ("asis", "cluster") or vector of environment levels in the order desired for plotting the heatmap. The default is "asis", where the rows and columns of the correlation matrix are maintained in environment level order. The "cluster" option invokes the agnes clustering algorithm on the FA correlation matrix then orders the rows and columns accordingly.

agnes.method

Clustering strategy; the default is "average".

Value

list

Details

The following function is taken from the ASExtras4 package. Please refer to this package and associated publications if you are interested in going deeper on this technique. You may be interested in reading and citing this publication if using this methodology

References

Cullis BR, Smith AB, Beek C, Cowling WA (2010). “Analysis of Yield and Oil from a Series of Canola Breeding Trials. Part II: Exploring VxE using Factor Analysis.” Genome, 53, 1002-1016. Smith AB, Cullis BR, Thompson R (2001). “Analyzing Variety by Environment Data Using Multiplicative Mixed Models and Adjustments for Spatial Field Trend.” Biometrics, 57, 1138-1147.

Examples

# library(tidyverse)
# library(asreml)
# library(agridat)
# data(besag.met)
# dat <- besag.met
#
# dat <- dat %>% arrange(county)
# model <- asreml(fixed = yield ~ 1 + county,
#                 random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
#                 residual = ~ dsum(~ units | county),
#                 data = dat,
#                 na.action = list(x="include",y="include"))
#
# ASM <- fa.asreml(model, trunc.char = NULL)