ggCor

Return a ggplot object to plot a triangular correlation figure between 2 or more variables.

ggCor(
  myData,
  colours = c("#db4437", "white", "#FF9D00"),
  blackLabs = c(-0.7, 0.7),
  showSignif = TRUE,
  pBreaks = c(0, 1e-04, 0.001, 0.01, Inf),
  pLabels = c("***", "**", "*", "ns"),
  showDiagonal = FALSE,
  Diag = NULL,
  returnTable = FALSE,
  returnN = FALSE
)

Arguments

myData
A data.frame with numerical columns for each variable to be compared.

colours
A vector of size three with the colors to be used for values -1, 0 and 1.

blackLabs
A numeric vector of size two, with min and max correlation coefficient

showSignif
Logical scalar. Display significance values ?

pBreaks
Passed to function 'cut'. Either a numeric vector of two or more unique cut points or a single number (greater than or equal to 2) giving the number of intervals into which x is to be cut.

pLabels
Passed to function 'cut'. labels for the levels of the resulting category. By default, labels are constructed using "(a,b]" interval notation. If pLabels = FALSE, simple integer codes are returned instead of a factor.

showDiagonal
Logical scalar. Display main diagonal values ?

Diag
A named vector of labels to display in the main diagonal. The names are used to place each value in the corresponding coordinates of the diagonal. Hence, these names must be the same as the colnames of myData

returnTable
Return the table to display instead of a ggplot object

returnN
Return plot with shared information

Value

A ggplot object containing a triangular correlation figure with all numeric variables in myData. If returnTable is TRUE, the table used to produce the figure is returned instead.

Author

Daniel Ariza

Examples

# ggCor(MM,colours = c("#db4437","white","#4285f4"), returnN = TRUE )