Heritability for Factor Analytic Models in ASReml-R

Heritability for Factor Analytic Models in ASReml-R

heritability_fa(
  model_fa = NULL,
  genotype = "line",
  env = "loc",
  vc.model = c("fa2"),
  diag = FALSE
)

Arguments

model_fa

Factor Analytic asreml model

genotype

A character string indicating the column in data that contains genotypes.

env

A character string indicating the column in data that contains environments or trials.

vc.model

A character string indicating the variance-covariance structure. Can be "fa1", "fa2", "fa3", "fa4" or "us".

diag

TRUE or FALSE depending on the user if they want to take the elements on the diagonal of the variance-covariance matrix or the elements out of the diagonal to estimate the heritability. FALSE by default.

Value

A list

Examples

# library(tidyverse)
# library(asreml)
# library(agridat)
# library(agriutilities)
# data(besag.met)
# dat <- besag.met
#
# dat <- dat %>% arrange(county)
# model <- asreml(
#   fixed = yield ~ 1 + county,
#   random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
#   residual = ~ dsum(~ units | county),
#   data = dat,
#   na.action = list(x="include",y="include")
# )
# heritability_fa(
#  model_fa = model,
#  genotype = "gen",
#  env = "county",
#  vc.model = "fa2"
# )