Multi-Environmental Trial Analysis

Multi-Environmental Trial Analysis

Usage

met_analysis(
  sma_output = NULL,
  h2_filter = 0.2,
  workspace = "1gb",
  vcov = NULL
)

Arguments

sma_output

Object of class smaAgri resulting of executing single_trial_analysis() function.

h2_filter

Numeric value to filter trials with poor heritability. 0.2 by default.

workspace

Sets the workspace for the core REML routines in the form of a number optionally followed directly by a valid measurement unit. "128mb" by default.

vcov

A character string specifying the Variance-Covariance structure to be fitted. Can be "fa2", "fa1", "us", or "corh". If NULL the function will try to fit an "us" Variance-Covariance and if it fails, it will try with "fa2" and then with "fa1".

Value

An object of class metAgri, with a list of:

trial_effects

A data.frame containing Trial BLUEs.

overall_BLUPs

A data.frame containing Genotypic BLUPs across trials, by trait.

BLUPs_GxE

A data.frame containing Genotypic BLUPs by trial/trait.

VCOV

A list by trait contanining the variance-covariance fitted.

stability

A data.frame containing several Stability coefficients resulting of executing the function stability().

heritability

A data.frame containing overall heritabilities by trait.

Examples

# \donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_MET(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
met_results <- met_analysis(out)
#> Online License checked out Thu Jan 12 10:53:32 2023
#> Online License checked out Thu Jan 12 10:53:33 2023
print(met_results)
#> ---------------------------------------------------------------------
#> Trial Effects (BLUEs):
#> ---------------------------------------------------------------------
#>   trait trial predicted.value std.error    status
#> 1 yield    C1       151.40258  1.351961 Estimable
#> 2 yield    C2        65.78758  1.119965 Estimable
#> 3 yield    C3        90.06743  1.478252 Estimable
#> 4 yield    C4       148.09868  1.241288 Estimable
#> 5 yield    C5       122.66271  1.429107 Estimable
#> 6 yield    C6        86.80613  1.492154 Estimable
#> 
#> ---------------------------------------------------------------------
#> Heritability:
#> ---------------------------------------------------------------------
#>   trait        h2
#> 1 yield 0.8244552
#> 
#> ---------------------------------------------------------------------
#> First Overall Predicted Values and Standard Errors (BLUPs):
#> ---------------------------------------------------------------------
#>   trait genotype predicted.value std.error    status
#> 1 yield      G01        110.0200  2.530462 Estimable
#> 2 yield      G02        111.1065  2.541857 Estimable
#> 3 yield      G03        102.7181  2.517908 Estimable
#> 4 yield      G04        115.5364  2.537029 Estimable
#> 5 yield      G05        120.2974  2.544593 Estimable
#> 6 yield      G06        109.4152  2.546403 Estimable
#> 
#> ---------------------------------------------------------------------
#> Variance-Covariance Matrix:
#> ---------------------------------------------------------------------
#> 
#> Correlation Matrix ('us'): yield
#>      C1   C2   C3   C4   C5   C6
#> C1 1.00 0.65 0.58 0.64 0.94 0.40
#> C2 0.65 1.00 0.57 0.79 0.59 0.74
#> C3 0.58 0.57 1.00 0.87 0.74 0.33
#> C4 0.64 0.79 0.87 1.00 0.66 0.41
#> C5 0.94 0.59 0.74 0.66 1.00 0.31
#> C6 0.40 0.74 0.33 0.41 0.31 1.00
#> 
#> Covariance Matrix ('us'): yield
#>       C1    C2    C3    C4     C5    C6
#> C1 74.02 27.52 46.28 32.19  81.34 28.63
#> C2 27.52 24.49 26.31 22.83  29.21 30.15
#> C3 46.28 26.31 86.39 47.71  69.12 25.64
#> C4 32.19 22.83 47.71 34.48  39.07 19.73
#> C5 81.34 29.21 69.12 39.07 101.57 25.94
#> C6 28.63 30.15 25.64 19.73  25.94 68.68
#> 
#> ---------------------------------------------------------------------
#> First Stability Coefficients:
#> ---------------------------------------------------------------------
#>   trait genotype superiority   static    wricke predicted.value
#> 1 yield      G57    22.49186 34.03947 15.764319        92.48602
#> 2 yield      G29    17.19476 34.67881  5.207048        99.16798
#> 3 yield      G59    17.07438 35.32410  4.801998        99.29619
#> 4 yield      G34    16.69884 34.35385  8.138613       100.06035
#> 5 yield      G10    15.76951 33.54824 11.883009       102.00207
#> 6 yield      G31    15.16212 32.21745 10.316367       102.40940
#> 
# }