Multi-Environmental Trial Analysis

Multi-Environmental Trial Analysis

met_analysis(
  sma_output = NULL,
  h2_filter = 0.2,
  workspace = "1gb",
  trials_to_fit_fa = 4
)

Arguments

sma_output

Object of class 'smaAgri' resulting of executing single_model_analysis function.

h2_filter

Numeric value to filter trials with poor heritability. 0.2 by default.

workspace

Sets the workspace for the core REML routines in the form of a number optionally followed directly by a valid measurement unit. "128mb" by default.

trials_to_fit_fa

Number of trials necessary to fit a Factor Analytic structure for the GxE interaction term. 4 by default.

Value

A list with a summary of the fitted models.

Examples

# \donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_MET(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_model_analysis(results, progress = FALSE)
out$resum_fitted_model
#>   trait trial heritability        CV    VarGen    VarErr      design
#> 1 yield    C1         0.72  6.057373  85.04747  84.80063 res_row_col
#> 2 yield    C2         0.34 17.906699  23.36095 110.89104 res_row_col
#> 3 yield    C3         0.63 12.189981  82.52294 119.34580 res_row_col
#> 4 yield    C4         0.38  8.222134  32.54677 142.90594 res_row_col
#> 5 yield    C5         0.80  6.805292 104.31185  62.88350 res_row_col
#> 6 yield    C6         0.51 15.267974  71.88271 178.58553 res_row_col
met_results <- met_analysis(out)
#> Online License checked out Mon Jan  9 19:30:35 2023
head(met_results$overall_BLUPs)
#>   trait genotype predicted.value std.error    status
#> 1 yield      G01        110.0104  2.531191 Estimable
#> 2 yield      G02        111.0692  2.542598 Estimable
#> 3 yield      G03        102.7300  2.518689 Estimable
#> 4 yield      G04        115.5712  2.537870 Estimable
#> 5 yield      G05        120.2970  2.544906 Estimable
#> 6 yield      G06        109.4588  2.546752 Estimable
# }