Single Trial Analysis

Single Trial Analysis

single_trial_analysis(
  results = NULL,
  progress = TRUE,
  engine = "asreml",
  remove_outliers = TRUE
)

Arguments

results

Object of class checkAgri resulting of executing check_design_MET() function.

progress

Should the progress of the modeling be printed. If TRUE, for every trial a line is output indicating the traits fitted for the particular trial.

engine

A character string specifying the name of the mixed modeling engine to use, either lme4 or asreml. For spatial designs, SpATS is always used, for other designs asreml as a default.

remove_outliers

Should outliers be removed? TRUE by default.

Value

A list of data.frames.

Examples

# \donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_MET(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
print(out)
#> ---------------------------------------------------------------------
#> Summary Fitted Models:
#> ---------------------------------------------------------------------
#>   trait trial heritability        CV    VarGen    VarErr      design
#> 1 yield    C1         0.72  6.057373  85.04747  84.80063 res_row_col
#> 2 yield    C2         0.34 17.906699  23.36095 110.89104 res_row_col
#> 3 yield    C3         0.63 12.189981  82.52294 119.34580 res_row_col
#> 4 yield    C4         0.38  8.222134  32.54677 142.90594 res_row_col
#> 5 yield    C5         0.80  6.805292 104.31185  62.88350 res_row_col
#> 6 yield    C6         0.51 15.267974  71.88271 178.58553 res_row_col
#> 
#> ---------------------------------------------------------------------
#> Outliers Removed:
#> ---------------------------------------------------------------------
#>   trait trial genotype   id outlier
#> 1 yield    C1      G60   50    TRUE
#> 2 yield    C3      G16  534    TRUE
#> 3 yield    C6      G08 1013    TRUE
#> 
#> ---------------------------------------------------------------------
#> First Predicted Values and Standard Errors (BLUEs/BLUPs):
#> ---------------------------------------------------------------------
#>   trait genotype trial    BLUEs  seBLUEs    BLUPs  seBLUPs         wt
#> 1 yield      G01    C1 142.0021 6.678480 144.3256 5.720102 0.02242047
#> 2 yield      G02    C1 160.7412 6.334549 157.4663 5.496182 0.02492118
#> 3 yield      G03    C1 130.8027 6.412041 135.3259 5.558685 0.02432245
#> 4 yield      G04    C1 155.7890 6.459256 154.2345 5.557676 0.02396818
#> 5 yield      G05    C1 163.7357 6.666347 161.2359 5.714016 0.02250216
#> 6 yield      G06    C1 128.2255 6.492415 133.6129 5.608807 0.02372398
#> 
# }