Stability Coefficients

Stability Coefficients

Usage

stability(
  predictions = NULL,
  genotype = NULL,
  trial = NULL,
  response = NULL,
  best = "max"
)

Arguments

predictions

A data.frame with one value per GxE combination.

genotype

A character string indicating the column in predictions that contains genotypes.

trial

A character string indicating the column in predictions that contains trials.

response

A character string specifying the response variable.

best

A character string specifying how to define the best genotype by numeric value ("min", "max"). "max" by default.

Value

A data.frame with several stability measures. "superiority" (cultivar-superiority measure), "static" (Shukla's stability variance) and "wricke" (Wricke's ecovalence).

Examples

# \donttest{
library(agridat)
library(agriutilities)

data(besag.met)
dat <- besag.met
results <- check_design_met(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
met_results <- met_analysis(out)

# Stability
stability(
  predictions = met_results$BLUPs_GxE,
  genotype = "genotype",
  trial = "trial",
  response = "predicted.value"
)
#>    genotype superiority   static    wricke predicted.value
#> 1       G57   22.675030 32.45871 13.962970        92.45997
#> 2       G29   17.275329 34.41794  4.343501        99.38429
#> 3       G34   17.262491 33.29276  8.514332        99.74688
#> 4       G59   16.948824 34.39425  4.798863        99.87221
#> 5       G31   16.230006 31.89042 11.722935       101.66382
#> 6       G10   15.752529 32.02994 11.499867       102.39802
#> 7       G03   15.641083 32.08671 12.686650       102.68118
#> 8       G33   15.141663 32.27229 11.588237       103.26889
#> 9       G12   14.759873 33.05361  6.796056       103.27584
#> 10      G37   13.963036 34.16955  4.327594       104.13181
#> 11      G50   13.641459 32.90413  7.070945       104.93693
#> 12      G23   13.003019 35.22175  3.864588       105.27946
#> 13      G11   13.105682 33.24260  7.886752       105.74274
#> 14      G18   12.517436 32.91550  6.885178       106.55068
#> 15      G28   11.958658 35.18371  2.419154       106.69629
#> 16      G58   12.135120 33.89965  3.648562       106.75031
#> 17      G44   12.077943 34.09435  3.995094       106.82398
#> 18      G21   11.830894 35.38075 10.379568       107.33195
#> 19      G47   12.090915 33.05583 11.817840       107.62747
#> 20      G39   11.173748 33.64909  4.014206       108.18333
#> 21      G26   10.735552 35.86332  3.971387       108.30097
#> 22      G06   11.517835 32.85864 12.892758       108.82966
#> 23      G43   10.966663 33.19977  7.343232       108.83672
#> 24      G41   10.406783 35.75248  4.182574       108.84645
#> 25      G24   10.544297 33.26269  5.041920       109.25773
#> 26      G60    9.741025 37.05767  8.780489       109.86799
#> 27      G07    9.762448 34.96258  2.179643       109.88117
#> 28      G01    9.937215 32.57294  6.679631       110.42974
#> 29      G02    9.771217 35.09288 13.201931       110.86168
#> 30      G56    8.864294 34.75210  6.002496       111.46879
#> 31      G49    9.009788 33.24373  5.617078       111.50788
#> 32      G64    9.046082 32.90382  5.604480       111.54649
#> 33      G35    8.384930 35.08863  3.248548       111.80537
#> 34      G15    8.136143 36.28056  6.763842       112.04034
#> 35      G63    8.698351 33.02662  8.366766       112.28387
#> 36      G16    9.333210 32.18294 13.596877       112.43129
#> 37      G46    7.989760 36.77815  9.730992       112.52801
#> 38      G19    7.824855 36.15393  7.239086       112.53976
#> 39      G14    8.151563 34.81858  7.926836       112.64328
#> 40      G25    7.752113 37.09972 10.356846       112.89019
#> 41      G53    7.620670 38.14092 12.457251       113.06116
#> 42      G62    7.972174 33.11630  6.708569       113.14709
#> 43      G13    8.175384 33.09835  9.025183       113.24000
#> 44      G20    7.039505 38.14772 13.208750       114.19955
#> 45      G27    7.294915 32.65496  6.171434       114.34265
#> 46      G40    6.186920 35.44688  4.606876       114.89417
#> 47      G42    6.085575 38.07851 12.536222       115.34878
#> 48      G04    6.076610 34.53664  5.406555       115.49459
#> 49      G22    5.769155 34.64815  2.136313       115.53529
#> 50      G09    6.014733 37.98175 13.021569       115.67568
#> 51      G48    5.737642 37.16551 10.150877       115.74334
#> 52      G17    6.054022 34.51968  9.067709       115.96178
#> 53      G36    6.025095 33.12237  5.322275       116.06844
#> 54      G51    5.369143 35.97648  7.161669       116.15439
#> 55      G45    4.952758 38.23459 13.684249       117.36414
#> 56      G30    4.877440 38.06062 13.126137       117.36716
#> 57      G38    5.003665 34.41899  9.628748       117.90802
#> 58      G52    3.826024 35.81902  5.849205       118.04168
#> 59      G55    4.608038 36.86081 12.256536       118.39797
#> 60      G54    4.105375 33.95087  3.472903       118.53412
#> 61      G32    3.315331 35.57753  5.767215       119.00367
#> 62      G08    3.084795 34.89594  4.161271       119.34622
#> 63      G61    3.488838 34.02658  4.996210       119.60318
#> 64      G05    2.604721 37.65734 12.045989       120.65996
# }